backend/requirements.txt
trueskill>=0.4.5
python-dotenv>=1.0.1
